Fix issues with edit post modal
- Couldn't click outside to close
- Double close button
- Loading Hylo on an edit URL for a request would break

Removes PostEditorModal and moves all functionality to CreateModal
