+++ src/components/CreateModal/index.js
@@ -1,3 +1,2 @@
 import CreateModal from './CreateModal'
-import connector from './CreateModal.connector'
-export default connector(CreateModal)
+export default CreateModal
