--- src/components/PostEditorModal/PostEditorModal.scss
@@ -1,90 +1,0 @@
-:global .post-editor-enter, :global .post-editor-appear {
-  .post-editor-background {
-    opacity: 0;
-  }
-
-  .post-editor-wrapper {
-    opacity: 0;
-    transform: translate(50%, -50%);
-  }
-}
-
-:global .post-editor-enter.post-editor-enter-active, :global .post-editor-appear.post-editor-appear-active {
-  .post-editor-background {
-    opacity: 1;
-    transition: opacity 200ms ease-out;
-  }
-
-  .post-editor-wrapper {
-    opacity: 1;
-    transform: translate(-50%, -50%);
-    transition: transform 300ms ease-out, opacity 400ms ease-out;
-  }
-}
-
-:global .post-editor-exit {
-  .post-editor-background {
-    opacity: 0;
-  }
-
-  .post-editor-wrapper {
-    opacity: 1;
-    transform: translate(0, 0);
-  }
-}
-
-:global .post-editor-exit.post-editor-exit-active {
-  .post-editor-background {
-    opacity: 0;
-    transition: opacity 300ms ease-out;
-  }
-
-  .post-editor-wrapper {
-    opacity: 0;
-    transform: translate(0, -100px);
-    transition: transform 200ms ease-out, opacity 200ms ease-out;
-  }
-}
-
-.post-editor-modal {
-  position: fixed;
-  z-index: 120;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  width: 100%;
-  height: 100%;
-  overflow-y: auto;
-  background: rgba(0, 0, 0, 0.4);
-}
-
-.post-editor-wrapper {
-    width: 100%;
-    max-width: 590px;
-    position: absolute;
-    z-index: 1;
-    top: 50%;
-    left: 50%;
-    margin: 0 0 0 0;
-    transform: translate(-50%, -50%);
-    max-height: 100vh;
-}
-
-.close-button {
-  cursor: pointer;
-  position: absolute;
-  top: 15px;
-  right: 15px;
-  z-index: 100;
-
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-@media screen and (max-width: 600px) {
-  .post-editor-wrapper {
-    width: 100%;
-  }
-}
