--- src/components/PostEditorModal/index.js
@@ -1,3 +1,0 @@
-import component from './PostEditorModal'
-
-export default component
